+++ HISTORY.rst
@@ -2,6 +2,7 @@
 ==================
 - Move IMAPMockServer to its own module.
 - Display feedback message when testing a mailing profile.
+- Add tests to the platform app.
 
 3.2.8 (2019-10-01)
 ==================